+++ src/main/java/dev/the_fireplace/caterpillar/client/screen/widget/TutorialButton.java
@@ -39,7 +39,7 @@ public class TutorialButton extends ImageButton {
             yOffset += this.yDiffTex;
         }
 
-        super.blit(poseStack, super.getX(), super.getY(), this.xTexStart, yOffset, super.width, super.height);
+        blit(poseStack, super.getX(), super.getY(), this.xTexStart, yOffset, super.width, super.height);
     }
 
     public void toggleTutorial() {